# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "assignment-5-customer-data",
    "version": "1.0.0",
    "description": "API for managing customers",
    "contact": {}
  },
  "servers": [],
  "paths": {
...
    "/customers/{id}/addresses": {
      "post": {
        "summary": "Create a new address for a specific customer",
        "description": "Create a new address and associate it with a specific customer using the customer's ID.",
        "operationId": "createAddress",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the customer to associate with the new address.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new address for the customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - all fields must be filled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields must be filled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all addresses for a specific customer",
        "description": "Retrieve a list of all addresses for a given customer.",
        "operationId": "getAllAddress",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the customer whose addresses are to be retrieved.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all addresses for the customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customerId}/addresses/{addressId}": {
      "get": {
        "summary": "Get address by ID",
        "description": "Retrieve a specific address by its ID, including the associated customer details.",
        "operationId": "getAddressById",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer associated with the address.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "addressId",
            "in": "path",
            "required": true,
            "description": "The ID of the address to retrieve.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the address by ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormattedAddress"
                }
              }
            }
          },
          "404": {
            "description": "Address not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/addresses/{addressId}": {
      "put": {
        "summary": "Update an existing address",
        "description": "Update the details of an existing address using the address ID.",
        "operationId": "updateAddress",
        "tags": ["Addresses"],
        "parameters": [
          {
            "name": "addressId",
            "in": "path",
            "required": true,
            "description": "The ID of the address to be updated.",
            "schema": {
              "type": "integer",
              "example": 101
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "adrs_street": {
                    "type": "string",
                    "description": "Street name of the address",
                    "example": "123 New Street"
                  },
                  "adrs_city": {
                    "type": "string",
                    "description": "City of the address",
                    "example": "Los Angeles"
                  },
                  "adrs_state": {
                    "type": "string",
                    "description": "State of the address",
                    "example": "CA"
                  },
                  "adrs_postalCode": {
                    "type": "string",
                    "description": "Postal code of the address",
                    "example": "90001"
                  }
                },
                "required": [
                  "adrs_street",
                  "adrs_city",
                  "adrs_state",
                  "adrs_postalCode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "previousDataAddress": {
                      "$ref": "#/components/schemas/Address"
                    },
                    "updatedAddress": {
                      "$ref": "#/components/schemas/Address"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - all fields must be filled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields must be filled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Address not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an existing address",
        "description": "Delete an existing address using the address ID.",
        "operationId": "deleteAddress",
        "tags": ["Addresses"],
        "parameters": [
          {
            "name": "addressId",
            "in": "path",
            "required": true,
            "description": "The ID of the address to be deleted.",
            "schema": {
              "type": "integer",
              "example": 101
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "404": {
            "description": "Address not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {
          "adrs_id": {
            "type": "integer",
            "description": "The unique ID of the address"
          },
          "adrs_cs_id": {
            "type": "integer",
            "description": "The ID of the customer associated with this address"
          },
          "adrs_street": {
            "type": "string",
            "description": "The street name of the address"
          },
          "adrs_city": {
            "type": "string",
            "description": "The city of the address"
          },
          "adrs_state": {
            "type": "string",
            "description": "The state of the address"
          },
          "adrs_postalCode": {
            "type": "string",
            "description": "The postal code of the address"
          }
        },
        "required": [
          "adrs_cs_id",
          "adrs_street",
          "adrs_city",
          "adrs_state",
          "adrs_postalCode"
        ]
      },
      "CreateAddressRequest": {
        "type": "object",
        "properties": {
          "adrs_street": {
            "type": "string",
            "description": "The street name of the address",
            "example": "123 Main St"
          },
          "adrs_city": {
            "type": "string",
            "description": "The city of the address",
            "example": "New York"
          },
          "adrs_state": {
            "type": "string",
            "description": "The state of the address",
            "example": "NY"
          },
          "adrs_postalCode": {
            "type": "string",
            "description": "The postal code of the address",
            "example": "10001"
          }
        },
        "required": [
          "adrs_street",
          "adrs_city",
          "adrs_state",
          "adrs_postalCode"
        ]
      },
...
      "FormattedAddress": {
        "type": "object",
        "properties": {
          "adrs_id": {
            "type": "integer",
            "description": "The unique ID of the address.",
            "example": 1
          },
          "adrs_street": {
            "type": "string",
            "description": "The street name of the address.",
            "example": "123 Main St"
          },
          "adrs_city": {
            "type": "string",
            "description": "The city of the address.",
            "example": "New York"
          },
          "adrs_state": {
            "type": "string",
            "description": "The state of the address.",
            "example": "NY"
          },
          "adrs_postalCode": {
            "type": "string",
            "description": "The postal code of the address.",
            "example": "10001"
          },
          "customer": {
            "type": "object",
            "description": "Details of the customer associated with the address.",
            "properties": {
              "cs_id": {
                "type": "integer",
                "description": "The unique ID of the customer.",
                "example": 1
              },
              "cs_fullName": {
                "type": "string",
                "description": "The full name of the customer.",
                "example": "Jane Doe"
              },
              "cs_email": {
                "type": "string",
                "description": "The email address of the customer.",
                "example": "jane.doe@example.com"
              },
              "cs_gender": {
                "type": "string",
                "enum": ["Male", "Female"],
                "description": "Gender of the customer.",
                "example": "Female"
              },
              "cs_birthDate": {
                "type": "string",
                "format": "date",
                "description": "Birth date of the customer in YYYY-MM-DD format.",
                "example": "1990-01-01"
              }
            },
            "nullable": true
          }
        }
      },
...
    }
  }
}